refactor: Fully migrate to VFS v2
- Removed the overlay_mount function from VfsManager as it is not yet implemented.
- Deleted overlayfs module and related tests to streamline the VFS structure.
- Updated imports in manager.rs and syscall.rs to reflect the new module organization.
- Added advanced tests for bind mounts, including nested and cross-VFS scenarios.
- Introduced performance tests for VFS v2 components, focusing on file creation, lookup, and directory operations.
- Enhanced test coverage for VFS manager creation and mount point functionality.
- Adjusted main.rs to use the new drivers module for initramfs.
