--- kernel/src/fs/drivers/cpio/tests/mkfs/initramfs/file1.txt
@@ -1,1 +1,0 @@
-Hello, world!
