--- kernel/src/fs/drivers/mod.rs
@@ -1,4 +1,0 @@
-pub mod cpio;
-pub mod initramfs;
-pub mod overlayfs;
-pub mod tmpfs;
