--- kernel/src/fs/drivers/overlayfs/tests.rs
@@ -1,1950 +1,0 @@
-//! Tests for OverlayFS implementation
-//! 
-//! These tests validate the OverlayFS functionality including layer management,
-//! file operations, and read-only behavior enforcement.
-
-use crate::fs::FileSystem;
-
-use super::OverlayFS;
-use super::*;
-use alloc::vec;
-use alloc::{string::String, vec::Vec};
-
-/// Test basic OverlayFS creation
-/// 
-/// This test validates that an OverlayFS can be created with no upper layer
-/// (read-only mode) and that it correctly identifies itself as "overlayfs".
-#[test_case]
-pub fn test_overlay_creation() {
-    // Test read-only overlay (no upper layer)
-    let lower_mount_nodes = Vec::new();
-    let lower_relative_paths = Vec::new();
-    
-    let overlay = OverlayFS::new(
-        None,
-        String::new(),
-        lower_mount_nodes,
-        lower_relative_paths,
-    );
-    
-    assert!(overlay.is_ok(), "Should be able to create read-only overlay");
-    
-    let overlay = overlay.unwrap();
-    assert_eq!(overlay.name(), "overlayfs", "Filesystem name should be 'overlayfs'");
-}
-
-/// Test OverlayFS creation with mismatched arrays
-/// 
-/// This test ensures that OverlayFS creation fails when the number of
-/// lower mount nodes doesn't match the number of lower relative paths.
-#[test_case]
-pub fn test_overlay_creation_mismatched_arrays() {
-    let lower_mount_nodes = Vec::new();
-    let lower_relative_paths = vec!["/test".to_string()];
-    
-    let overlay = OverlayFS::new(
-        None,
-        String::new(),
-        lower_mount_nodes,
-        lower_relative_paths,
-    );
-    
-    assert!(overlay.is_err(), "Should fail with mismatched array lengths");
-}
-
-/// Test that read-only overlay rejects write operations
-/// 
-/// This test verifies that an OverlayFS without an upper layer correctly
-/// rejects write operations like file creation, directory creation, and removal.
-#[test_case]
-pub fn test_readonly_overlay_rejects_writes() {
-    let overlay = OverlayFS::new(
-        None, // No upper layer = read-only
-        String::new(),
-        Vec::new(),
-        Vec::new(),
-    ).unwrap();
-    
-    // These operations should fail because there's no upper layer
-    let create_result = overlay.create_file("/test.txt", FileType::RegularFile);
-    assert!(create_result.is_err(), "Should fail to create file on read-only overlay");
-    
-    let mkdir_result = overlay.create_dir("/test_dir");
-    assert!(mkdir_result.is_err(), "Should fail to create directory on read-only overlay");
-    
-    let remove_result = overlay.remove("/test.txt");
-    assert!(remove_result.is_err(), "Should fail to remove file on read-only overlay");
-}
-
-/// Test write flag detection for COW trigger
-/// 
-/// This test verifies that the OverlayFS correctly detects write operations
-/// from file open flags and would trigger copy-up for files in lower layers.
-#[test_case]
-pub fn test_write_flag_detection() {
-    let overlay = OverlayFS::new(
-        None, // No upper layer for simplicity
-        String::new(),
-        Vec::new(),
-        Vec::new(),
-    ).unwrap();
-    
-    // Test various flag combinations that should trigger write detection
-    
-    // O_WRONLY (1) - write-only
-    let result_wronly = overlay.open("/test.txt", 1);
-    assert!(result_wronly.is_err(), "Should fail on read-only overlay with write flag");
-    
-    // O_RDWR (2) - read/write  
-    let result_rdwr = overlay.open("/test.txt", 2);
-    assert!(result_rdwr.is_err(), "Should fail on read-only overlay with read/write flag");
-    
-    // O_RDONLY (0) - read-only should not trigger write detection
-    let result_rdonly = overlay.open("/test.txt", 0);
-    // This will fail because file doesn't exist, but not because of write detection
-    if let Err(e) = result_rdonly {
-        // Should fail with "File not found" not "Cannot write to read-only overlay"
-        assert!(e.message.contains("File not found"), "Should fail with file not found, not write permission error");
-    }
-    
-    // O_WRONLY | O_APPEND (1 | 1024 = 1025) - append mode
-    let result_append = overlay.open("/test.txt", 1025);
-    assert!(result_append.is_err(), "Should fail on read-only overlay with append flag");
-    
-    // O_RDWR | O_APPEND (2 | 1024 = 1026) - read/write append mode
-    let result_rdwr_append = overlay.open("/test.txt", 1026);
-    assert!(result_rdwr_append.is_err(), "Should fail on read-only overlay with read/write append flag");
-}
-
-/// Test COW logic components individually
-/// 
-/// This test verifies that the OverlayFS correctly handles file operations
-/// when no filesystems are present, which is useful for testing boundary conditions.
-#[test_case]
-pub fn test_cow_logic_components() {
-    let overlay = OverlayFS::new(
-        None, // No upper layer
-        String::new(),
-        Vec::new(),
-        Vec::new(),
-    ).unwrap();
-    
-    // Test that operations fail appropriately when no filesystems are available
-    
-    // metadata should fail when no layers exist
-    let metadata_result = overlay.metadata("/any_file.txt");
-    assert!(metadata_result.is_err(), "Should fail when no layers exist");
-    
-    // readdir should also fail when no layers exist (consistent with other operations)
-    let readdir_result = overlay.readdir("/");
-    assert!(readdir_result.is_err(), "Should fail when no layers exist");
-    
-    // The internal helper functions are private, but their effects are testable
-    // through the public interface. The logic ensures that:
-    // - Files are searched in upper layer first, then lower layers
-    // - COW is triggered only when writing to lower-only files  
-    // - All operations fail when no layers exist (consistent behavior)
-}
-
-/// Test Copy-on-Write (COW) behavior during file writes using overlay_mount
-/// 
-/// This test verifies that when opening a file for writing that exists only
-/// in the lower layer, the OverlayFS correctly performs copy-up to the upper layer.
-/// It also verifies that the written data is correctly stored and read back.
-#[test_case]
-pub fn test_cow_on_file_write_with_overlay_mount() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    // Create VFS manager
-    let manager = VfsManager::new();
-    
-    // Create and register upper TmpFS (1MB limit)
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    // Create and register lower TmpFS with initial content
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_file("/test_file.txt", FileType::RegularFile);
-
-    let lower_content = b"original content";
-    
-    // Write initial content to lower filesystem
-    if let Ok(lower_file) = lower_fs.open("/test_file.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(lower_content);
-    }
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount both filesystems
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    // Create overlay mount: upper=/upper, lower=/lower, target=/overlay
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec!["/lower"],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Phase 1: Verify initial state
-            let upper_file_check = manager.open("/upper/test_file.txt", 0); // O_RDONLY
-            assert!(upper_file_check.is_err(), "File should not exist in upper layer initially");
-            
-            // Verify file exists in lower layer
-            assert!(manager.open("/lower/test_file.txt", 0).is_ok(), "File should exist in lower layer");
-            
-            // Phase 2: Write through overlay (should trigger COW)
-            let open_result = manager.open("/overlay/test_file.txt", 1); // O_WRONLY
-            
-            match open_result {
-                Ok(kernel_obj) => {
-                    if let Some(stream_ops) = kernel_obj.as_stream() {
-                        // Write specific new content through overlay
-                        let test_content = b"COW modified content";
-                        let write_result = stream_ops.write(test_content);
-                        assert!(write_result.is_ok(), "Should be able to write to overlay file");
-                        assert_eq!(write_result.unwrap(), test_content.len(), "Should write all bytes");
-                        
-                        // Phase 3: Verify COW occurred - file should now exist in upper layer
-                        let upper_file_after = manager.open("/upper/test_file.txt", 0); // O_RDONLY
-                        assert!(upper_file_after.is_ok(), "File should exist in upper layer after COW");
-                        
-                        // Phase 4: Verify upper layer has the written content (COW worked)
-                        if let Ok(upper_kernel_obj) = upper_file_after {
-                            if let Some(upper_stream) = upper_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 128];
-                                if let Ok(bytes_read) = upper_stream.read(&mut buffer) {
-                                    assert!(bytes_read > 0, "Upper layer should have some content after COW");
-                                    let content = &buffer[..bytes_read];
-                                    assert_eq!(content, test_content, "Upper layer should contain the written content");
-                                }
-                            }
-                        }
-                        
-                        // Phase 5: Verify overlay reads the written content from upper layer
-                        if let Ok(overlay_kernel_obj) = manager.open("/overlay/test_file.txt", 0) {
-                            if let Some(overlay_stream) = overlay_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 128];
-                                if let Ok(bytes_read) = overlay_stream.read(&mut buffer) {
-                                    assert!(bytes_read > 0, "Overlay should read some content from upper layer");
-                                    let content = &buffer[..bytes_read];
-                                    assert_eq!(content, test_content, "Overlay should read the written content from upper layer");
-                                }
-                            }
-                        }
-                        
-                        // Phase 6: Verify lower layer is unchanged
-                        if let Ok(lower_kernel_obj) = manager.open("/lower/test_file.txt", 0) {
-                            if let Some(lower_stream) = lower_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 128];
-                                if let Ok(bytes_read) = lower_stream.read(&mut buffer) {
-                                    let content = &buffer[..bytes_read];
-                                    assert_eq!(content, lower_content, "Lower layer should remain unchanged after COW");
-                                }
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    assert!(false, "COW write operation should succeed: {}", e.message);
-                }
-            }
-        }
-        Err(e) => {
-            assert!(false, "Overlay mount should succeed: {}", e.message);
-        }
-    }
-}
-
-/// Test Copy-on-Write (COW) behavior during file append operations
-/// 
-/// This test verifies that append operations (O_APPEND flag) also trigger COW
-/// when the file exists only in the lower layer. We focus on COW behavior
-/// rather than exact content verification due to append implementation details.
-#[test_case]
-pub fn test_cow_on_file_append_with_overlay_mount() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    // Create VFS manager
-    let manager = VfsManager::new();
-    
-    // Create and register upper TmpFS (1MB limit)
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    // Create and register lower TmpFS with test file and initial content
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_file("/append_test.txt", FileType::RegularFile);
-    
-    // Write initial content to lower filesystem
-    if let Ok(lower_file) = lower_fs.open("/append_test.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(b"original content");
-    }
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount both filesystems
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    // Create overlay mount
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec!["/lower"],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Phase 1: Verify initial state
-            let upper_file_check = manager.open("/upper/append_test.txt", 0); // O_RDONLY
-            assert!(upper_file_check.is_err(), "File should not exist in upper layer initially");
-            
-            // Verify file exists in lower layer
-            assert!(manager.open("/lower/append_test.txt", 0).is_ok(), "File should exist in lower layer");
-            
-            // Verify file is visible through overlay
-            assert!(manager.open("/overlay/append_test.txt", 0).is_ok(), "File should be visible through overlay");
-            
-            // Phase 2: Append through overlay (should trigger COW)
-            // O_WRONLY | O_APPEND = 1 | 1024 = 1025
-            let open_result = manager.open("/overlay/append_test.txt", 1025);
-            
-            match open_result {
-                Ok(kernel_obj) => {
-                    if let Some(stream_ops) = kernel_obj.as_stream() {
-                        // Append some data through overlay
-                        let append_data = b" + appended data";
-                        let append_result = stream_ops.write(append_data);
-                        assert!(append_result.is_ok(), "Should be able to append to overlay file");
-                        assert_eq!(append_result.unwrap(), append_data.len(), "Should write all appended bytes");
-                        
-                        // Phase 3: Verify COW occurred - file should now exist in upper layer
-                        let upper_file_after = manager.open("/upper/append_test.txt", 0); // O_RDONLY
-                        assert!(upper_file_after.is_ok(), "File should exist in upper layer after COW");
-                        
-                        // Phase 4: Verify that overlay now reads from upper layer (some content should exist)
-                        if let Ok(overlay_kernel_obj) = manager.open("/overlay/append_test.txt", 0) {
-                            if let Some(overlay_stream) = overlay_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 128];
-                                if let Ok(bytes_read) = overlay_stream.read(&mut buffer) {
-                                    assert!(bytes_read > 0, "Overlay should read some content from upper layer");
-                                }
-                            }
-                        }
-                        
-                        // Phase 5: Verify upper layer has some content (COW worked)
-                        if let Ok(upper_kernel_obj) = upper_file_after {
-                            if let Some(upper_stream) = upper_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 128];
-                                if let Ok(bytes_read) = upper_stream.read(&mut buffer) {
-                                    assert!(bytes_read > 0, "Upper layer should have some content after COW");
-                                }
-                            }
-                        }
-                        
-                        // Phase 6: Verify lower layer remains unchanged (should still be accessible)
-                        assert!(manager.open("/lower/append_test.txt", 0).is_ok(), "Lower layer file should remain accessible after COW");
-                    }
-                }
-                Err(e) => {
-                    assert!(false, "COW append operation should succeed: {}", e.message);
-                }
-            }
-        }
-        Err(e) => {
-            assert!(false, "Overlay mount should succeed: {}", e.message);
-        }
-    }
-}
-
-/// Test that read-only overlay mount works correctly
-/// 
-/// This test verifies that overlay mounts without upper layer correctly
-/// reject write operations while allowing read operations.
-#[test_case]
-pub fn test_readonly_overlay_mount() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    // Create VFS manager
-    let manager = VfsManager::new();
-    
-    // Create and register lower TmpFS with test file
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_file("/readonly_test.txt", FileType::RegularFile);
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount lower filesystem
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    // Create read-only overlay mount (no upper layer)
-    let overlay_result = manager.overlay_mount(
-        None, // No upper layer = read-only
-        vec!["/lower"],
-        "/readonly_overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Test: Read operations should work
-            let read_result = manager.open("/readonly_overlay/readonly_test.txt", 0); // O_RDONLY
-            assert!(read_result.is_ok(), "Read operations should work on read-only overlay");
-            
-            // Test: Write operations should fail
-            let write_result = manager.open("/readonly_overlay/readonly_test.txt", 1); // O_WRONLY
-            
-            match write_result {
-                Ok(_) => {
-                    // Should not succeed on read-only overlay
-                    assert!(false, "Write operation should not succeed on read-only overlay");
-                }
-                Err(e) => {
-                    // Expected: should fail with permission denied or similar
-                    assert!(e.message.contains("read-only") || e.message.contains("permission") || e.message.contains("Cannot write"),
-                           "Should fail with appropriate read-only error message: {}", e.message);
-                }
-            }
-        }
-        Err(_) => {
-            // Expected in some test environments
-        }
-    }
-}
-
-/// Test detailed COW verification - ensures file is actually copied to upper layer
-/// 
-/// This test performs comprehensive verification of the COW mechanism including:
-/// - File existence checks before and after COW
-/// - Write operation success with exact data verification
-/// - Upper layer file creation with correct content
-/// - Content isolation between upper and lower layers
-#[test_case]
-pub fn test_detailed_cow_verification() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    // Create VFS manager
-    let manager = VfsManager::new();
-    
-    // Create and register upper TmpFS (1MB limit)
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    // Create and register lower TmpFS with initial content
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_file("/detail_test.txt", FileType::RegularFile);
-    
-    // Write specific content to lower filesystem
-    if let Ok(lower_file) = lower_fs.open("/detail_test.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(b"initial lower data");
-    }
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount both filesystems
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    // Create overlay mount
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec!["/lower"],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Phase 1: Verify initial state and content
-            assert!(manager.open("/upper/detail_test.txt", 0).is_err(), 
-                   "File should not exist in upper layer initially");
-            assert!(manager.open("/lower/detail_test.txt", 0).is_ok(), 
-                   "File should exist in lower layer");
-            assert!(manager.open("/overlay/detail_test.txt", 0).is_ok(), 
-                   "File should be visible through overlay");
-            
-            // Verify initial content from overlay (should read from lower)
-            if let Ok(overlay_read) = manager.open("/overlay/detail_test.txt", 0) {
-                if let Some(overlay_stream) = overlay_read.as_stream() {
-                    let mut buffer = [0u8; 64];
-                    if let Ok(bytes_read) = overlay_stream.read(&mut buffer) {
-                        let content = &buffer[..bytes_read];
-                        assert_eq!(content, b"initial lower data", "Overlay should initially read from lower layer");
-                    }
-                }
-            }
-            
-            // Phase 2: Perform write operation (should trigger COW)
-            let write_result = manager.open("/overlay/detail_test.txt", 1); // O_WRONLY
-            
-            match write_result {
-                Ok(kernel_obj) => {
-                    if let Some(stream_ops) = kernel_obj.as_stream() {
-                        // Write specific test content
-                        let test_data = b"COW test data 12345";
-                        let write_op = stream_ops.write(test_data);
-                        assert!(write_op.is_ok(), "Write operation should succeed");
-                        assert_eq!(write_op.unwrap(), test_data.len(), "Should write all bytes");
-                        
-                        // Phase 3: Verify COW occurred - file should now exist in upper layer
-                        let upper_check = manager.open("/upper/detail_test.txt", 0);
-                        assert!(upper_check.is_ok(), 
-                               "File should exist in upper layer after write (COW should have occurred)");
-                        
-                        // Phase 4: Verify upper layer has some content (COW worked)
-                        if let Ok(upper_kernel_obj) = upper_check {
-                            if let Some(upper_stream) = upper_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 64];
-                                if let Ok(bytes_read) = upper_stream.read(&mut buffer) {
-                                    assert!(bytes_read > 0, "Upper layer should have some content after COW");
-                                }
-                            }
-                        }
-                        
-                        // Phase 5: Verify overlay now reads from upper layer (some content exists)
-                        if let Ok(read_kernel_obj) = manager.open("/overlay/detail_test.txt", 0) {
-                            if let Some(read_stream) = read_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 64];
-                                if let Ok(bytes_read) = read_stream.read(&mut buffer) {
-                                    assert!(bytes_read > 0, "Overlay should read some content from upper layer");
-                                }
-                            }
-                        }
-                        
-                        // Phase 6: Verify lower layer is completely unchanged
-                        if let Ok(lower_kernel_obj) = manager.open("/lower/detail_test.txt", 0) {
-                            if let Some(lower_stream) = lower_kernel_obj.as_stream() {
-                                let mut buffer = [0u8; 64];
-                                if let Ok(bytes_read) = lower_stream.read(&mut buffer) {
-                                    let content = &buffer[..bytes_read];
-                                    assert_eq!(content, b"initial lower data", "Lower layer should remain completely unchanged");
-                                }
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    assert!(false, "COW test failed - write operation failed: {}", e.message);
-                }
-            }
-        }
-        Err(e) => {
-            assert!(false, "Overlay mount failed: {}", e.message);
-        }
-    }
-}
-
-/// Test COW behavior with multiple write operations
-/// 
-/// Verifies that multiple writes to the same file work correctly after COW
-/// and that each write operation produces the expected content changes.
-#[test_case]
-pub fn test_cow_multiple_writes() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Setup filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_file("/multi_write.txt", FileType::RegularFile);
-    
-    // Write initial content to lower layer
-    if let Ok(lower_file) = lower_fs.open("/multi_write.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(b"base content");
-    }
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Verify initial content exists through overlay (should read from lower)
-        assert!(manager.open("/overlay/multi_write.txt", 0).is_ok(), "Should be able to read from overlay initially");
-        
-        // First write (triggers COW)
-        let first_data = b"first write data";
-        if let Ok(file1) = manager.open("/overlay/multi_write.txt", 1) {
-            if let Some(stream1) = file1.as_stream() {
-                let write_result = stream1.write(first_data);
-                assert!(write_result.is_ok(), "First write should succeed");
-                assert_eq!(write_result.unwrap(), first_data.len(), "Should write all bytes");
-            }
-        }
-        
-        // Verify file exists in upper after first write
-        assert!(manager.open("/upper/multi_write.txt", 0).is_ok(), 
-               "File should exist in upper after first write");
-        
-        // Verify first write content is readable through overlay
-        if let Ok(read_after_first) = manager.open("/overlay/multi_write.txt", 0) {
-            if let Some(read_stream) = read_after_first.as_stream() {
-                let mut buffer = [0u8; 128];
-                if let Ok(bytes_read) = read_stream.read(&mut buffer) {
-                    assert!(bytes_read > 0, "Should read some content after first write");
-                    let content = &buffer[..bytes_read];
-                    assert_eq!(content, first_data, "Should read the first written content");
-                }
-            }
-        }
-        
-        // Second write (should work on upper layer file)
-        let second_data = b"second write content";
-        if let Ok(file2) = manager.open("/overlay/multi_write.txt", 1) {
-            if let Some(stream2) = file2.as_stream() {
-                let write_result = stream2.write(second_data);
-                assert!(write_result.is_ok(), "Second write should succeed");
-                assert_eq!(write_result.unwrap(), second_data.len(), "Should write all bytes");
-            }
-        }
-        
-        // Verify second write content is readable through overlay
-        if let Ok(read_after_second) = manager.open("/overlay/multi_write.txt", 0) {
-            if let Some(read_stream) = read_after_second.as_stream() {
-                let mut buffer = [0u8; 128];
-                if let Ok(bytes_read) = read_stream.read(&mut buffer) {
-                    assert!(bytes_read > 0, "Should read some content after second write");
-                    let content = &buffer[..bytes_read];
-                    assert_eq!(content, second_data, "Should read the second written content");
-                }
-            }
-        }
-        
-        // Verify upper layer contains the second write content
-        if let Ok(upper_read) = manager.open("/upper/multi_write.txt", 0) {
-            if let Some(upper_stream) = upper_read.as_stream() {
-                let mut buffer = [0u8; 128];
-                if let Ok(bytes_read) = upper_stream.read(&mut buffer) {
-                    assert!(bytes_read > 0, "Upper layer should have some content");
-                    let content = &buffer[..bytes_read];
-                    assert_eq!(content, second_data, "Upper layer should contain the second written content");
-                }
-            }
-        }
-        
-        // Verify lower layer remains accessible
-        assert!(manager.open("/lower/multi_write.txt", 0).is_ok(), "Lower layer should remain accessible");
-    }
-}
-
-/// Test that COW correctly handles write operations
-/// 
-/// This test verifies that COW triggers correctly for both write and append
-/// operations without checking specific file content details.
-#[test_case]
-pub fn test_cow_overwrite_vs_preserve() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Setup filesystems with two test files
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_file("/write_test.txt", FileType::RegularFile);
-    let _ = lower_fs.create_file("/append_test.txt", FileType::RegularFile);
-    
-    // Write initial content to both files
-    if let Ok(file1) = lower_fs.open("/write_test.txt", 1) {
-        let _ = file1.write(b"initial content 1");
-    }
-    if let Ok(file2) = lower_fs.open("/append_test.txt", 1) {
-        let _ = file2.write(b"initial content 2");
-    }
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Test 1: Write operation (triggers COW)
-        let new_content = b"new write content";
-        if let Ok(write_file) = manager.open("/overlay/write_test.txt", 1) { // O_WRONLY
-            if let Some(stream) = write_file.as_stream() {
-                let write_result = stream.write(new_content);
-                assert!(write_result.is_ok(), "Write should succeed");
-                assert_eq!(write_result.unwrap(), new_content.len(), "Should write all bytes");
-            }
-        }
-        
-        // Test 2: Append operation (also triggers COW)
-        if let Ok(append_file) = manager.open("/overlay/append_test.txt", 1025) { // O_WRONLY | O_APPEND
-            if let Some(stream) = append_file.as_stream() {
-                let append_result = stream.write(b"appended data");
-                assert!(append_result.is_ok(), "Append should succeed");
-            }
-        }
-        
-        // Verify both files exist in upper layer after COW
-        assert!(manager.open("/upper/write_test.txt", 0).is_ok(), 
-               "Write file should exist in upper layer after COW");
-        assert!(manager.open("/upper/append_test.txt", 0).is_ok(), 
-               "Append file should exist in upper layer after COW");
-        
-        // Verify write file content through overlay (normal write should work)
-        if let Ok(overlay_write_file) = manager.open("/overlay/write_test.txt", 0) {
-            if let Some(stream) = overlay_write_file.as_stream() {
-                let mut buffer = [0u8; 128];
-                if let Ok(bytes_read) = stream.read(&mut buffer) {
-                    let content = &buffer[..bytes_read];
-                    assert_eq!(content, new_content, "Overlay should read the written content");
-                }
-            }
-        }
-        
-        // Verify append file exists but don't check content (append not fully implemented)
-        assert!(manager.open("/overlay/append_test.txt", 0).is_ok(), 
-               "Append file should be readable through overlay");
-        
-        // Verify write file content in upper layer
-        if let Ok(upper_write_file) = manager.open("/upper/write_test.txt", 0) {
-            if let Some(stream) = upper_write_file.as_stream() {
-                let mut buffer = [0u8; 128];
-                if let Ok(bytes_read) = stream.read(&mut buffer) {
-                    let content = &buffer[..bytes_read];
-                    assert_eq!(content, new_content, "Upper layer should contain the written content");
-                }
-            }
-        }
-        
-        // Verify lower layer remains accessible
-        assert!(manager.open("/lower/write_test.txt", 0).is_ok(), 
-               "Lower write file should remain accessible");
-        assert!(manager.open("/lower/append_test.txt", 0).is_ok(), 
-               "Lower append file should remain accessible");
-    }
-}
-
-/// Test basic COW functionality without content verification
-/// 
-/// This test verifies that COW correctly moves files from lower to upper layer
-/// during write operations, without checking specific file content details.
-#[test_case]
-pub fn test_cow_basic_behavior() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    
-    // Create a file with some data
-    let _ = lower_fs.create_file("/data_test.txt", FileType::RegularFile);
-    if let Ok(lower_file) = lower_fs.open("/data_test.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(b"some data");
-    }
-    
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    let _ = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay");
-    
-    // Phase 1: Verify initial state
-    assert!(manager.open("/overlay/data_test.txt", 0).is_ok(), "File should be visible through overlay");
-    assert!(manager.open("/lower/data_test.txt", 0).is_ok(), "File should exist in lower layer");
-    assert!(manager.open("/upper/data_test.txt", 0).is_err(), "File should not exist in upper layer initially");
-    
-    // Phase 2: Write data through overlay (trigger COW)
-    let test_content = b"new data from overlay";
-    if let Ok(overlay_file) = manager.open("/overlay/data_test.txt", 1) { // O_WRONLY
-        if let Some(stream) = overlay_file.as_stream() {
-            // Write specific data to trigger COW
-            let write_result = stream.write(test_content);
-            assert!(write_result.is_ok(), "Should be able to write through overlay");
-            assert_eq!(write_result.unwrap(), test_content.len(), "Should write all bytes");
-        }
-    }
-    
-    // Phase 3: Verify COW occurred - file should now exist in upper layer
-    assert!(manager.open("/upper/data_test.txt", 0).is_ok(), "File should exist in upper layer after COW");
-    
-    // Phase 4: Verify overlay reads the written content from upper layer
-    if let Ok(overlay_file) = manager.open("/overlay/data_test.txt", 0) { // O_RDONLY
-        if let Some(stream) = overlay_file.as_stream() {
-            let mut buffer = [0u8; 100];
-            if let Ok(bytes_read) = stream.read(&mut buffer) {
-                assert!(bytes_read > 0, "Overlay should read some content after COW");
-                let content = &buffer[..bytes_read];
-                assert_eq!(content, test_content, "Overlay should read the written content");
-            }
-        }
-    }
-    
-    // Phase 5: Verify upper layer contains the written content
-    if let Ok(upper_file) = manager.open("/upper/data_test.txt", 0) { // O_RDONLY
-        if let Some(stream) = upper_file.as_stream() {
-            let mut buffer = [0u8; 100];
-            if let Ok(bytes_read) = stream.read(&mut buffer) {
-                assert!(bytes_read > 0, "Upper layer should have some content after COW");
-                let content = &buffer[..bytes_read];
-                assert_eq!(content, test_content, "Upper layer should contain the written content");
-            }
-        }
-    } else {
-        assert!(false, "File should exist in upper layer after COW");
-    }
-    
-    // Phase 6: Verify lower layer remains accessible
-    assert!(manager.open("/lower/data_test.txt", 0).is_ok(), "Lower layer should remain accessible after COW");
-}
-
-/// Test basic file creation and write operations with content verification
-/// 
-/// This test verifies that new files created through overlay correctly store
-/// the written content and can be read back accurately.
-#[test_case]
-pub fn test_overlay_file_write_content() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Test 1: Create a new file through overlay and write content
-        if let Ok(()) = manager.create_file("/overlay/new_file.txt", FileType::RegularFile) {
-            if let Ok(file_obj) = manager.open("/overlay/new_file.txt", 1) { // O_WRONLY
-                if let Some(stream) = file_obj.as_stream() {
-                    let test_data = b"Hello, OverlayFS!";
-                    let write_result = stream.write(test_data);
-                    assert!(write_result.is_ok(), "Should be able to write to new file");
-                    assert_eq!(write_result.unwrap(), test_data.len(), "Should write all bytes");
-                    
-                    // Verify file exists in upper layer
-                    assert!(manager.open("/upper/new_file.txt", 0).is_ok(), 
-                           "New file should exist in upper layer");
-                    
-                    // Read back content through overlay
-                    if let Ok(read_obj) = manager.open("/overlay/new_file.txt", 0) { // O_RDONLY
-                        if let Some(read_stream) = read_obj.as_stream() {
-                            let mut buffer = [0u8; 32];
-                            if let Ok(bytes_read) = read_stream.read(&mut buffer) {
-                                let content = &buffer[..bytes_read];
-                                assert_eq!(content, test_data, "Should read back the exact written content");
-                            }
-                        }
-                    }
-                    
-                    // Read back content directly from upper layer
-                    if let Ok(upper_obj) = manager.open("/upper/new_file.txt", 0) { // O_RDONLY
-                        if let Some(upper_stream) = upper_obj.as_stream() {
-                            let mut buffer = [0u8; 32];
-                            if let Ok(bytes_read) = upper_stream.read(&mut buffer) {
-                                let content = &buffer[..bytes_read];
-                                assert_eq!(content, test_data, "Upper layer should contain the exact written content");
-                            }
-                        }
-                    }
-                    
-                    // Test truncate functionality on the new file
-                    let truncate_result = manager.truncate("/overlay/new_file.txt", 5);
-                    assert!(truncate_result.is_ok(), "Should be able to truncate new file");
-                    
-                    // Verify truncated size and content
-                    if let Ok(metadata) = manager.metadata("/overlay/new_file.txt") {
-                        assert_eq!(metadata.size, 5, "File size should be 5 bytes after truncation");
-                    }
-                    
-                    if let Ok(truncated_obj) = manager.open("/overlay/new_file.txt", 0) { // O_RDONLY
-                        if let Some(truncated_stream) = truncated_obj.as_stream() {
-                            let mut buffer = [0u8; 10];
-                            if let Ok(bytes_read) = truncated_stream.read(&mut buffer) {
-                                assert_eq!(bytes_read, 5, "Should read exactly 5 bytes after truncation");
-                                let content = &buffer[..bytes_read];
-                                assert_eq!(content, b"Hello", "Content should be truncated to first 5 characters");
-                            }
-                        }
-                    }
-                    
-                    // Test expanding the file again
-                    let expand_result = manager.truncate("/overlay/new_file.txt", 12);
-                    assert!(expand_result.is_ok(), "Should be able to expand file");
-                    
-                    if let Ok(metadata) = manager.metadata("/overlay/new_file.txt") {
-                        assert_eq!(metadata.size, 12, "File size should be 12 bytes after expansion");
-                    }
-                    
-                    // Verify file does not exist in lower layer
-                    assert!(manager.open("/lower/new_file.txt", 0).is_err(), 
-                           "New file should not exist in lower layer");
-                }
-            }
-        }
-    }
-}
-
-/// Test overwrite operation with content verification
-/// 
-/// This test verifies that when a file is overwritten (seek to start + write),
-/// the content is correctly replaced and readable.
-#[test_case]
-pub fn test_overlay_file_overwrite_content() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_content = b"original content";
-    
-    // Create a file in lower layer with initial content
-    let _ = lower_fs.create_file("/test_overwrite.txt", FileType::RegularFile);
-    if let Ok(lower_file) = lower_fs.open("/test_overwrite.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(lower_content);
-    }
-    
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Verify initial content through overlay
-        if let Ok(initial_obj) = manager.open("/overlay/test_overwrite.txt", 0) { // O_RDONLY
-            if let Some(initial_stream) = initial_obj.as_stream() {
-                let mut buffer = [0u8; 64];
-                if let Ok(bytes_read) = initial_stream.read(&mut buffer) {
-                    let content = &buffer[..bytes_read];
-                    assert_eq!(content, lower_content, "Should initially read from lower layer");
-                }
-            }
-        }
-        
-        // Overwrite file through overlay (triggers COW)
-        if let Ok(write_obj) = manager.open("/overlay/test_overwrite.txt", 1) { // O_WRONLY
-            if let Some(write_stream) = write_obj.as_stream() {
-                let new_content = b"completely new content";
-                let write_result = write_stream.write(new_content);
-                assert!(write_result.is_ok(), "Should be able to overwrite file");
-                assert_eq!(write_result.unwrap(), new_content.len(), "Should write all bytes");
-                
-                // Verify file exists in upper layer after COW
-                assert!(manager.open("/upper/test_overwrite.txt", 0).is_ok(), 
-                       "File should exist in upper layer after overwrite");
-                
-                // Read back new content through overlay
-                if let Ok(read_obj) = manager.open("/overlay/test_overwrite.txt", 0) { // O_RDONLY
-                    if let Some(read_stream) = read_obj.as_stream() {
-                        let mut buffer = [0u8; 64];
-                        if let Ok(bytes_read) = read_stream.read(&mut buffer) {
-                            let content = &buffer[..bytes_read];
-                            assert_eq!(content, new_content, "Overlay should read the new content after overwrite");
-                        }
-                    }
-                }
-                
-                // Verify upper layer contains new content
-                if let Ok(upper_obj) = manager.open("/upper/test_overwrite.txt", 0) { // O_RDONLY
-                    if let Some(upper_stream) = upper_obj.as_stream() {
-                        let mut buffer = [0u8; 64];
-                        if let Ok(bytes_read) = upper_stream.read(&mut buffer) {
-                            let content = &buffer[..bytes_read];
-                            assert_eq!(content, new_content, "Upper layer should contain the new content");
-                        }
-                    }
-                }
-                
-                // Verify lower layer remains unchanged
-                if let Ok(lower_obj) = manager.open("/lower/test_overwrite.txt", 0) { // O_RDONLY
-                    if let Some(lower_stream) = lower_obj.as_stream() {
-                        let mut buffer = [0u8; 64];
-                        if let Ok(bytes_read) = lower_stream.read(&mut buffer) {
-                            let content = &buffer[..bytes_read];
-                            assert_eq!(content, lower_content, "Lower layer should remain unchanged");
-                        }
-                    }
-                }
-                
-                // Test truncate functionality on the overwritten file
-                let truncate_result = manager.truncate("/overlay/test_overwrite.txt", 10);
-                assert!(truncate_result.is_ok(), "Should be able to truncate overwritten file");
-                
-                // Verify truncated size and content
-                if let Ok(metadata) = manager.metadata("/overlay/test_overwrite.txt") {
-                    assert_eq!(metadata.size, 10, "File size should be 10 bytes after truncation");
-                }
-                
-                if let Ok(truncated_obj) = manager.open("/overlay/test_overwrite.txt", 0) { // O_RDONLY
-                    if let Some(truncated_stream) = truncated_obj.as_stream() {
-                        let mut buffer = [0u8; 32];
-                        if let Ok(bytes_read) = truncated_stream.read(&mut buffer) {
-                            assert_eq!(bytes_read, 10, "Should read exactly 10 bytes after truncation");
-                            let content = &buffer[..bytes_read];
-                            let content_str = core::str::from_utf8(content).unwrap_or("");
-                            assert_eq!(content_str, "completely", "Content should be truncated to first 10 characters");
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Debug test to investigate write behavior
-// #[test_case]
-pub fn test_debug_write_behavior() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_content = b"original content";
-    
-    // Create a file in lower layer with initial content
-    let _ = lower_fs.create_file("/debug_test.txt", FileType::RegularFile);
-    if let Ok(lower_file) = lower_fs.open("/debug_test.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(lower_content);
-    }
-    
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Write through overlay
-        if let Ok(write_obj) = manager.open("/overlay/debug_test.txt", 1) { // O_WRONLY
-            if let Some(write_stream) = write_obj.as_stream() {
-                let new_content = b"new";
-                let write_result = write_stream.write(new_content);
-                assert!(write_result.is_ok(), "Should be able to write");
-                assert_eq!(write_result.unwrap(), new_content.len(), "Should write all bytes");
-                
-                // Read back and print the content length and bytes
-                if let Ok(read_obj) = manager.open("/overlay/debug_test.txt", 0) { // O_RDONLY
-                    if let Some(read_stream) = read_obj.as_stream() {
-                        let mut buffer = [0u8; 64];
-                        if let Ok(bytes_read) = read_stream.read(&mut buffer) {
-                            // Print debug info
-                            crate::println!("DEBUG: bytes_read = {}", bytes_read);
-                            crate::println!("DEBUG: content = {:?}", &buffer[..bytes_read]);
-                            crate::println!("DEBUG: expected = {:?}", new_content);
-                            crate::println!("INFO : If the filesystem is not implemented truncating, this test may not work as expected.");
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Test whiteout functionality
-/// 
-/// This test verifies that files can be hidden from lower layers using whiteout files
-/// when they are removed through the overlay.
-#[test_case]
-pub fn test_whiteout_basic() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    
-    // Create a file in lower layer only
-    let _ = lower_fs.create_file("/hidden_file.txt", FileType::RegularFile);
-    if let Ok(lower_file) = lower_fs.open("/hidden_file.txt", 1) { // O_WRONLY
-        let _ = lower_file.write(b"file in lower layer");
-    }
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Phase 1: Verify file is visible through overlay initially
-        assert!(manager.open("/overlay/hidden_file.txt", 0).is_ok(), 
-               "File should be visible through overlay initially");
-        assert!(manager.open("/lower/hidden_file.txt", 0).is_ok(), 
-               "File should exist in lower layer");
-        assert!(manager.open("/upper/hidden_file.txt", 0).is_err(), 
-               "File should not exist in upper layer initially");
-        
-        // Phase 2: Remove file through overlay (should create whiteout)
-        let remove_result = manager.remove("/overlay/hidden_file.txt");
-        assert!(remove_result.is_ok(), "Should be able to remove file through overlay");
-        
-        // Phase 3: Verify file is no longer visible through overlay
-        assert!(manager.open("/overlay/hidden_file.txt", 0).is_err(), 
-               "File should be hidden after removal");
-        
-        // Phase 4: Verify file still exists in lower layer
-        assert!(manager.open("/lower/hidden_file.txt", 0).is_ok(), 
-               "File should still exist in lower layer");
-        
-        // Phase 5: Verify whiteout file exists in upper layer
-        // Whiteout files are named ".wh.<original_name>"
-        assert!(manager.open("/upper/.wh.hidden_file.txt", 0).is_ok(), 
-               "Whiteout file should exist in upper layer");
-    }
-}
-
-/// Test whiteout with directory listing
-/// 
-/// This test verifies that removed files don't appear in directory listings
-/// even though they still exist in lower layers.
-#[test_case]
-pub fn test_whiteout_directory_listing() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    
-    // Create multiple files in lower layer
-    let _ = lower_fs.create_file("/visible.txt", FileType::RegularFile);
-    let _ = lower_fs.create_file("/to_be_hidden.txt", FileType::RegularFile);
-    let _ = lower_fs.create_file("/also_visible.txt", FileType::RegularFile);
-    
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Phase 1: Verify all files are visible initially (+ . and ..)
-        if let Ok(entries) = manager.readdir("/overlay/") {
-            assert_eq!(entries.len(), 5, "Should see all 5 entries (3 files + . and ..)");
-            let names: Vec<&str> = entries.iter().map(|e| e.name.as_str()).collect();
-            assert!(names.contains(&"visible.txt"));
-            assert!(names.contains(&"to_be_hidden.txt"));
-            assert!(names.contains(&"also_visible.txt"));
-            assert!(names.contains(&"."));
-            assert!(names.contains(&".."));
-        }
-        
-        // Phase 2: Remove one file (should create whiteout)
-        let remove_result = manager.remove("/overlay/to_be_hidden.txt");
-        assert!(remove_result.is_ok(), "Should be able to remove file");
-        
-        // Phase 3: Verify directory listing excludes removed file (4 entries: 2 files + . and ..)
-        if let Ok(entries) = manager.readdir("/overlay/") {
-            assert_eq!(entries.len(), 4, "Should see only 4 entries after removal (2 files + . and ..)");
-            let names: Vec<&str> = entries.iter().map(|e| e.name.as_str()).collect();
-            assert!(names.contains(&"visible.txt"));
-            assert!(!names.contains(&"to_be_hidden.txt"), "Removed file should not appear");
-            assert!(names.contains(&"also_visible.txt"));
-        }
-        
-        // Phase 4: Verify lower layer still has all files (+ . and ..)
-        if let Ok(entries) = manager.readdir("/lower/") {
-            assert_eq!(entries.len(), 5, "Lower layer should still have all 5 entries (3 files + . and ..)");
-        }
-        
-        // Phase 5: Verify whiteout doesn't appear in overlay listing
-        if let Ok(entries) = manager.readdir("/overlay/") {
-            let names: Vec<&str> = entries.iter().map(|e| e.name.as_str()).collect();
-            assert!(!names.iter().any(|name| name.starts_with(".wh.")), 
-                   "Whiteout files should not appear in overlay listing");
-        }
-    }
-}
-
-/// Test removal of file that exists in upper layer
-/// 
-/// This test verifies that files existing in the upper layer are simply
-/// removed without creating whiteout files.
-#[test_case]
-pub fn test_remove_upper_layer_file() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Phase 1: Create file directly in upper layer through overlay
-        let create_result = manager.create_regular_file("/overlay/upper_file.txt");
-        assert!(create_result.is_ok(), "Should be able to create file in overlay");
-        
-        // Verify file exists in upper layer
-        assert!(manager.open("/upper/upper_file.txt", 0).is_ok(), 
-               "File should exist in upper layer");
-        assert!(manager.open("/overlay/upper_file.txt", 0).is_ok(), 
-               "File should be visible through overlay");
-        
-        // Phase 2: Remove file (should just delete from upper, no whiteout needed)
-        let remove_result = manager.remove("/overlay/upper_file.txt");
-        assert!(remove_result.is_ok(), "Should be able to remove file");
-        
-        // Phase 3: Verify file is completely gone
-        assert!(manager.open("/upper/upper_file.txt", 0).is_err(), 
-               "File should not exist in upper layer after removal");
-        assert!(manager.open("/overlay/upper_file.txt", 0).is_err(), 
-               "File should not be visible through overlay after removal");
-        
-        // Phase 4: Verify no whiteout file was created
-        assert!(manager.open("/upper/.wh.upper_file.txt", 0).is_err(), 
-               "No whiteout file should be created for upper layer file");
-    }
-}
-
-/// Test attempting to remove non-existent file
-/// 
-/// This test verifies that removing a file that doesn't exist in any layer
-/// returns an appropriate error.
-#[test_case]
-pub fn test_remove_nonexistent_file() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Attempt to remove non-existent file
-        let remove_result = manager.remove("/overlay/nonexistent.txt");
-        assert!(remove_result.is_err(), "Should fail to remove non-existent file");
-        
-        // Verify error is appropriate (file not found)
-        if let Err(e) = remove_result {
-            assert_eq!(e.kind, FileSystemErrorKind::NotFound, 
-                      "Should return NotFound error for non-existent file");
-        }
-    }
-}
-
-/// Test removing and recreating file with same name
-/// 
-/// This test verifies that after removing a file from the lower layer (creating whiteout)
-/// and then creating a new file with the same name, the new file is visible and the
-/// lower layer file remains hidden. This tests proper whiteout handling when files
-/// are recreated.
-#[test_case]
-pub fn test_remove_and_recreate_same_name() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Phase 1: Create file in lower layer
-        let lower_create_result = manager.create_regular_file("/lower/test_file.txt");
-        assert!(lower_create_result.is_ok(), "Should be able to create file in lower layer");
-        
-        // Write original content to lower layer file
-        if let Ok(kernel_obj) = manager.open("/lower/test_file.txt", 1) { // O_WRONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let original_content = b"original content in lower layer";
-                let _ = stream_ops.write(original_content);
-            }
-        }
-        
-        // Verify file is visible through overlay
-        assert!(manager.open("/overlay/test_file.txt", 0).is_ok(), 
-               "File should be visible through overlay initially");
-        
-        // Phase 2: Remove file through overlay (creates whiteout)
-        let remove_result = manager.remove("/overlay/test_file.txt");
-        assert!(remove_result.is_ok(), "Should be able to remove file through overlay");
-        
-        // Verify file is no longer visible through overlay
-        assert!(manager.open("/overlay/test_file.txt", 0).is_err(), 
-               "File should not be visible through overlay after removal");
-
-        // Verify file is removed from upper layer
-        assert!(manager.open("/upper/test_file.txt", 0).is_err(), 
-               "File should not exist in upper layer after removal");
-        
-        // Verify lower layer file still exists
-        assert!(manager.open("/lower/test_file.txt", 0).is_ok(), 
-               "Lower layer file should still exist");
-        
-        // Verify whiteout file exists in upper layer
-        assert!(manager.open("/upper/.wh.test_file.txt", 0).is_ok(), 
-               "Whiteout file should exist in upper layer");
-        
-        // Phase 3: Create new file with same name through overlay
-        let new_create_result = manager.create_regular_file("/overlay/test_file.txt");
-        match new_create_result {
-            Err(ref e) => {
-                crate::println!("DEBUG: Error creating new file: {:?}", e);
-            },
-            Ok(_) => {}
-        }
-        assert!(new_create_result.is_ok(), "Should be able to create new file with same name");
-        
-        // Write new content to the new file
-        if let Ok(kernel_obj) = manager.open("/overlay/test_file.txt", 1) { // O_WRONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let new_content = b"new content in upper layer";
-                let _ = stream_ops.write(new_content);
-            }
-        }
-        
-        // Phase 4: Verify new file is visible through overlay
-        assert!(manager.open("/overlay/test_file.txt", 0).is_ok(), 
-               "New file should be visible through overlay");
-        
-        // Verify new file exists in upper layer
-        assert!(manager.open("/upper/test_file.txt", 0).is_ok(), 
-               "New file should exist in upper layer");
-        
-        // Phase 5: Verify content is from new file, not lower layer
-        if let Ok(kernel_obj) = manager.open("/overlay/test_file.txt", 0) { // O_RDONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let mut buffer = [0u8; 64];
-                if let Ok(bytes_read) = stream_ops.read(&mut buffer) {
-                    let content = &buffer[..bytes_read];
-                    let content_str = core::str::from_utf8(content).unwrap_or("");
-                    // Check that we're reading from the new upper layer file
-                    assert!(content_str.contains("new content"), 
-                           "Should read new content from upper layer, got: {}", content_str);
-                    assert!(!content_str.contains("original content"), 
-                           "Should not read original content from lower layer");
-                }
-            }
-        }
-        
-        // Phase 6: Verify lower layer file still contains original content
-        if let Ok(kernel_obj) = manager.open("/lower/test_file.txt", 0) { // O_RDONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let mut buffer = [0u8; 64];
-                if let Ok(bytes_read) = stream_ops.read(&mut buffer) {
-                    let content = &buffer[..bytes_read];
-                    let content_str = core::str::from_utf8(content).unwrap_or("");
-                    assert!(content_str.contains("original content"), 
-                           "Lower layer should still contain original content");
-                }
-            }
-        }
-        
-        // Phase 7: Verify directory listing shows only the new file
-        if let Ok(entries) = manager.readdir("/overlay/") {
-            let names: Vec<&str> = entries.iter().map(|e| e.name.as_str()).collect();
-            assert!(names.contains(&"test_file.txt"), "New file should appear in directory listing");
-            
-            // Count how many times the filename appears (should be exactly once)
-            let count = names.iter().filter(|&&name| name == "test_file.txt").count();
-            assert_eq!(count, 1, "File should appear exactly once in directory listing");
-            
-            // Verify whiteout doesn't appear in listing
-            assert!(!names.iter().any(|name| name.starts_with(".wh.")), 
-                   "Whiteout files should not appear in overlay listing");
-        }
-        
-        // Phase 8: Verify whiteout file should be removed/ignored after recreation
-        // When a new file is created with the same name, the whiteout should no longer affect visibility
-        // The exact behavior may vary (whiteout might still exist but be ignored, or be removed)
-        // What's important is that the new file is visible and functional
-        
-        // Additional verification: ensure we can still read/write the new file
-        // Note: We'll just verify the file is writable, but won't append due to the complex
-        // interaction with lack of truncate support that can cause confusing test results
-        if let Ok(kernel_obj) = manager.open("/overlay/test_file.txt", 1) { // O_WRONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                // Just verify the file is writable
-                let verify_content = b"verified";
-                let _ = stream_ops.write(verify_content);
-            }
-        }
-        
-        // Read back and verify we can still read from the new file
-        if let Ok(kernel_obj) = manager.open("/overlay/test_file.txt", 0) { // O_RDONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let mut buffer = [0u8; 128];
-                if let Ok(bytes_read) = stream_ops.read(&mut buffer) {
-                    let content = &buffer[..bytes_read];
-                    let content_str = core::str::from_utf8(content).unwrap_or("");
-                    assert!(content_str.starts_with("verified"), 
-                           "Should read back the verification content, got: {}", content_str);
-                }
-            }
-        }
-    }
-}
-
-/// Test file truncation functionality in OverlayFS
-/// 
-/// This test verifies that truncate operations work correctly with OverlayFS,
-/// including COW behavior when truncating files that exist only in lower layers.
-#[test_case]
-pub fn test_overlay_truncate_functionality() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Phase 1: Create file in lower layer with initial content
-        let _ = manager.create_file("/lower/test_truncate.txt", FileType::RegularFile);
-        
-        // Write initial content to lower layer (100 bytes)
-        if let Ok(kernel_obj) = manager.open("/lower/test_truncate.txt", 1) { // O_WRONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let initial_content = b"This is a long test content that will be used for truncation testing in the overlay filesystem......";
-                assert_eq!(initial_content.len(), 100, "Initial content should be exactly 100 bytes");
-                let _ = stream_ops.write(initial_content);
-            }
-        }
-        
-        // Verify initial file size through overlay
-        if let Ok(metadata) = manager.metadata("/overlay/test_truncate.txt") {
-            assert_eq!(metadata.size, 100, "Initial file size should be 100 bytes");
-        }
-        
-        // Phase 2: Truncate file to smaller size (triggers COW)
-        let truncate_result = manager.truncate("/overlay/test_truncate.txt", 50);
-        assert!(truncate_result.is_ok(), "Should be able to truncate file through overlay");
-        
-        // Verify file was copied to upper layer and truncated
-        assert!(manager.open("/upper/test_truncate.txt", 0).is_ok(), 
-               "File should exist in upper layer after truncate");
-        
-        // Verify new file size
-        if let Ok(metadata) = manager.metadata("/overlay/test_truncate.txt") {
-            assert_eq!(metadata.size, 50, "File size should be 50 bytes after truncation");
-        }
-        
-        // Verify truncated content
-        if let Ok(kernel_obj) = manager.open("/overlay/test_truncate.txt", 0) { // O_RDONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let mut buffer = [0u8; 100];
-                if let Ok(bytes_read) = stream_ops.read(&mut buffer) {
-                    assert_eq!(bytes_read, 50, "Should read exactly 50 bytes");
-                    let content = &buffer[..bytes_read];
-                    let content_str = core::str::from_utf8(content).unwrap_or("");
-                    assert_eq!(content_str, "This is a long test content that will be used for ", 
-                              "Content should be truncated at 50 characters");
-                }
-            }
-        }
-        
-        // Phase 3: Expand file (truncate to larger size)
-        let expand_result = manager.truncate("/overlay/test_truncate.txt", 80);
-        assert!(expand_result.is_ok(), "Should be able to expand file through overlay");
-        
-        // Verify expanded file size
-        if let Ok(metadata) = manager.metadata("/overlay/test_truncate.txt") {
-            assert_eq!(metadata.size, 80, "File size should be 80 bytes after expansion");
-        }
-        
-        // Verify expanded content (should be padded with zeros)
-        if let Ok(kernel_obj) = manager.open("/overlay/test_truncate.txt", 0) { // O_RDONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let mut buffer = [0u8; 100];
-                if let Ok(bytes_read) = stream_ops.read(&mut buffer) {
-                    assert_eq!(bytes_read, 80, "Should read exactly 80 bytes");
-                    let content = &buffer[..bytes_read];
-                    
-                    // First 50 bytes should be original content
-                    let first_part = &content[..50];
-                    let first_str = core::str::from_utf8(first_part).unwrap_or("");
-                    assert_eq!(first_str, "This is a long test content that will be used for ",
-                              "First 50 bytes should be original content");
-                    
-                    // Remaining 30 bytes should be zeros
-                    let padding = &content[50..];
-                    assert!(padding.iter().all(|&b| b == 0), 
-                           "Padding bytes should all be zeros");
-                }
-            }
-        }
-        
-        // Phase 4: Truncate to zero (empty file)
-        let empty_result = manager.truncate("/overlay/test_truncate.txt", 0);
-        assert!(empty_result.is_ok(), "Should be able to truncate file to zero");
-        
-        // Verify empty file
-        if let Ok(metadata) = manager.metadata("/overlay/test_truncate.txt") {
-            assert_eq!(metadata.size, 0, "File size should be 0 bytes after truncation to zero");
-        }
-        
-        // Verify lower layer file remains unchanged
-        if let Ok(metadata) = manager.metadata("/lower/test_truncate.txt") {
-            assert_eq!(metadata.size, 100, "Lower layer file should remain 100 bytes");
-        }
-    }
-}
-
-/// Test truncation of file that exists in upper layer
-/// 
-/// This test verifies that truncation works correctly for files that were
-/// created directly in the upper layer (no COW needed).
-#[test_case]
-pub fn test_overlay_truncate_upper_layer_file() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Create file directly in overlay (goes to upper layer)
-        let _ = manager.create_file("/overlay/upper_file.txt", FileType::RegularFile);
-        
-        // Write content to upper layer file
-        if let Ok(kernel_obj) = manager.open("/overlay/upper_file.txt", 1) { // O_WRONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let content = b"This file exists only in upper layer and will be truncated.";
-                let _ = stream_ops.write(content);
-            }
-        }
-        
-        // Verify initial file size
-        if let Ok(metadata) = manager.metadata("/overlay/upper_file.txt") {
-            assert!(metadata.size > 0, "Initial file size should be greater than 0");
-        }
-        
-        // Truncate file
-        let truncate_result = manager.truncate("/overlay/upper_file.txt", 30);
-        assert!(truncate_result.is_ok(), "Should be able to truncate upper layer file");
-        
-        // Verify truncated size
-        if let Ok(metadata) = manager.metadata("/overlay/upper_file.txt") {
-            assert_eq!(metadata.size, 30, "File size should be 30 bytes after truncation");
-        }
-        
-        // Verify file still exists only in upper layer
-        assert!(manager.open("/upper/upper_file.txt", 0).is_ok(), 
-               "File should still exist in upper layer");
-        assert!(manager.open("/lower/upper_file.txt", 0).is_err(), 
-               "File should not exist in lower layer");
-    }
-}
-
-/// Test truncation error handling
-/// 
-/// This test verifies that truncate operations return appropriate errors
-/// for non-existent files and read-only overlays.
-#[test_case]
-pub fn test_overlay_truncate_error_handling() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Test 1: Truncate on read-only overlay
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(None, vec!["/lower"], "/readonly_overlay") {
-        // Try to truncate on read-only overlay
-        let truncate_result = manager.truncate("/readonly_overlay/nonexistent.txt", 100);
-        assert!(truncate_result.is_err(), "Should fail to truncate on read-only overlay");
-    }
-    
-    // Test 2: Truncate non-existent file on writable overlay
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    let _ = manager.mount(upper_fs_id, "/upper");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Try to truncate non-existent file
-        let truncate_result = manager.truncate("/overlay/does_not_exist.txt", 100);
-        assert!(truncate_result.is_err(), "Should fail to truncate non-existent file");
-    }
-}
-
-/// Test truncation with COW and whiteout interactions
-/// 
-/// This test verifies complex scenarios involving truncation, COW, and whiteout files.
-#[test_case]
-pub fn test_overlay_truncate_with_whiteout() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create filesystems
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    if let Ok(()) = manager.overlay_mount(Some("/upper"), vec!["/lower"], "/overlay") {
-        // Phase 1: Create file in lower layer
-        let _ = manager.create_file("/lower/whiteout_test.txt", FileType::RegularFile);
-        if let Ok(kernel_obj) = manager.open("/lower/whiteout_test.txt", 1) { // O_WRONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let content = b"This file will be removed and then recreated for truncation test.";
-                let _ = stream_ops.write(content);
-            }
-        }
-        
-        // Phase 2: Remove file through overlay (creates whiteout)
-        let remove_result = manager.remove("/overlay/whiteout_test.txt");
-        assert!(remove_result.is_ok(), "Should be able to remove file through overlay");
-        
-        // Verify file is hidden
-        assert!(manager.open("/overlay/whiteout_test.txt", 0).is_err(), 
-               "File should be hidden after removal");
-        
-        // Phase 3: Recreate file with same name
-        let _ = manager.create_file("/overlay/whiteout_test.txt", FileType::RegularFile);
-        if let Ok(kernel_obj) = manager.open("/overlay/whiteout_test.txt", 1) { // O_WRONLY
-           
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let content = b"This is the new file content after recreation and will be truncated.";
-                let _ = stream_ops.write(content);
-            }
-        }
-        
-        // Phase 4: Truncate the recreated file
-        let truncate_result = manager.truncate("/overlay/whiteout_test.txt", 25);
-        assert!(truncate_result.is_ok(), "Should be able to truncate recreated file");
-        
-        // Verify truncated content
-        if let Ok(metadata) = manager.metadata("/overlay/whiteout_test.txt") {
-            assert_eq!(metadata.size, 25, "File size should be 25 bytes after truncation");
-        }
-        
-        if let Ok(kernel_obj) = manager.open("/overlay/whiteout_test.txt", 0) { // O_RDONLY
-            if let Some(stream_ops) = kernel_obj.as_stream() {
-                let mut buffer = [0u8; 50];
-                if let Ok(bytes_read) = stream_ops.read(&mut buffer) {
-                    assert_eq!(bytes_read, 25, "Should read exactly 25 bytes");
-                    let content = &buffer[..bytes_read];
-                    let content_str = core::str::from_utf8(content).unwrap_or("");
-                    assert_eq!(content_str, "This is the new file cont",
-                              "Content should be truncated at 25 characters");
-                }
-            }
-        }
-        
-        // Verify lower layer file still exists but is hidden
-        assert!(manager.open("/lower/whiteout_test.txt", 0).is_ok(), 
-               "Lower layer file should still exist");
-    }
-}
-
-/// Test OverlayFS directory object functionality
-/// 
-/// This test validates that OverlayDirectoryObject correctly merges entries
-/// from multiple layers when a directory is opened and read.
-#[test_case]
-pub fn test_overlay_directory_object() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    // Create VFS manager
-    let manager = VfsManager::new();
-    
-    // Create and register upper TmpFS
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    upper_fs.create_dir("/test_dir").expect("Failed to create directory '/test_dir'");
-    let _ = upper_fs.create_file("/test_dir/upper_file.txt", FileType::RegularFile);
-    let _ = upper_fs.create_file("/test_dir/lower_file1.txt", FileType::RegularFile); // Override lower
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    // Create and register lower TmpFS
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    lower_fs.create_dir("/test_dir").expect("Failed to create directory '/test_dir'");
-    let _ = lower_fs.create_file("/test_dir/lower_file1.txt", FileType::RegularFile);
-    let _ = lower_fs.create_file("/test_dir/lower_file2.txt", FileType::RegularFile);
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount both filesystems
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    // Create overlay mount
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec!["/lower"],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Open the overlay directory
-            let dir_obj = manager.open("/overlay/test_dir", 0)
-                .expect("Failed to open overlay directory");
-            
-            // Get FileObject from KernelObject
-            if let Some(file) = dir_obj.as_file() {
-                // Read all directory entries
-                let mut buffer = [0u8; 512]; // Increased buffer size for DirectoryEntry
-                let mut entries_read = 0;
-                
-                loop {
-                    match file.read(&mut buffer) {
-                        Ok(0) => break, // EOF
-                        Ok(_) => entries_read += 1,
-                        Err(e) => panic!("Failed to read directory entry: {:?}", e),
-                    }
-                }
-                
-                // Should read 5 entries: ".", "..", upper_file.txt, lower_file1.txt (from upper), lower_file2.txt
-                assert_eq!(entries_read, 5, "Should read exactly 5 directory entries (including . and ..)");
-            } else {
-                panic!("Expected FileObject from directory open");
-            }
-        }
-        Err(e) => panic!("Failed to create overlay mount: {:?}", e),
-    }
-}
-
-/// Test OverlayFS directory object seek operations
-/// 
-/// This test validates that seek operations work correctly on directory objects.
-#[test_case]
-pub fn test_overlay_directory_seek() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create upper and lower filesystems with test files
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = upper_fs.create_dir("/test_dir");
-    let _ = upper_fs.create_file("/test_dir/file1.txt", FileType::RegularFile);
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_dir("/test_dir");
-    let _ = lower_fs.create_file("/test_dir/file2.txt", FileType::RegularFile);
-    let _ = lower_fs.create_file("/test_dir/file3.txt", FileType::RegularFile);
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec!["/lower"],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Open directory
-            let dir_obj = manager.open("/overlay/test_dir", 0)
-                .expect("Failed to open directory");
-            
-            if let Some(file) = dir_obj.as_file() {
-                // Test seek operations (now 5 entries: ., .., file1.txt, file2.txt, file3.txt)
-                assert_eq!(file.seek(SeekFrom::Start(1)).unwrap(), 1, "Seek to start position 1");
-                assert_eq!(file.seek(SeekFrom::Current(1)).unwrap(), 2, "Seek current +1");
-                assert_eq!(file.seek(SeekFrom::End(-1)).unwrap(), 4, "Seek to end-1 (5 entries)");
-            } else {
-                panic!("Expected FileObject from directory open");
-            }
-        }
-        Err(e) => panic!("Failed to create overlay mount: {:?}", e),
-    }
-}
-
-/// Test OverlayFS directory object metadata
-/// 
-/// This test validates that metadata for directory objects is correct.
-#[test_case]
-pub fn test_overlay_directory_metadata() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = upper_fs.create_dir("/test_dir");
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    let _ = manager.mount(upper_fs_id, "/upper");
-    
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec![],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // Open directory
-            let dir_obj = manager.open("/overlay/test_dir", 0)
-                .expect("Failed to open directory");
-            
-            if let Some(file) = dir_obj.as_file() {
-                // Check metadata
-                let metadata = file.metadata().expect("Failed to get metadata");
-                
-                assert_eq!(metadata.file_type, FileType::Directory, "Should be directory type");
-                assert!(metadata.permissions.read, "Directory should be readable");
-                assert!(metadata.permissions.execute, "Directory should be executable for traversal");
-            } else {
-                panic!("Expected FileObject from directory open");
-            }
-        }
-        Err(e) => panic!("Failed to create overlay mount: {:?}", e),
-    }
-}
-
-/// Test OverlayFS directory object with whiteout files
-/// 
-/// This test validates that whiteout files properly hide files from lower layers
-/// when reading directory contents via the directory object.
-#[test_case]
-pub fn test_overlay_directory_whiteout() {
-    use crate::fs::{VfsManager, drivers::tmpfs::TmpFS, FileType};
-    use alloc::boxed::Box;
-    
-    let manager = VfsManager::new();
-    
-    // Create upper filesystem with whiteout
-    let upper_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = upper_fs.create_dir("/test_dir");
-    let _ = upper_fs.create_file("/test_dir/.wh.hidden_file.txt", FileType::RegularFile);
-    let upper_fs_id = manager.register_fs(upper_fs);
-    
-    // Create lower filesystem with files
-    let lower_fs = Box::new(TmpFS::new(1024 * 1024));
-    let _ = lower_fs.create_dir("/test_dir");
-    let _ = lower_fs.create_file("/test_dir/hidden_file.txt", FileType::RegularFile);
-    let _ = lower_fs.create_file("/test_dir/visible_file.txt", FileType::RegularFile);
-    let lower_fs_id = manager.register_fs(lower_fs);
-    
-    // Mount and create overlay
-    let _ = manager.mount(upper_fs_id, "/upper");
-    let _ = manager.mount(lower_fs_id, "/lower");
-    
-    let overlay_result = manager.overlay_mount(
-        Some("/upper"),
-        vec!["/lower"],
-        "/overlay"
-    );
-    
-    match overlay_result {
-        Ok(()) => {
-            // First, test readdir directly to see what entries are present
-            let direct_entries = manager.readdir("/overlay/test_dir")
-                .expect("Failed to readdir overlay directory");
-            
-            // Should see 3 entries: ".", "..", visible_file.txt (hidden_file.txt should be hidden by whiteout)
-            assert_eq!(direct_entries.len(), 3, "readdir should return 3 entries (including . and ..)");
-            
-            // Check that only visible_file.txt is among the regular files (excluding . and ..)
-            let regular_files: Vec<_> = direct_entries.iter()
-                .filter(|e| !e.name.starts_with('.') || (e.name != "." && e.name != ".."))
-                .collect();
-            assert_eq!(regular_files.len(), 1, "Should only see 1 regular file");
-            assert_eq!(regular_files[0].name, "visible_file.txt", "Should only see visible_file.txt");
-            
-            // Open directory
-            let dir_obj = manager.open("/overlay/test_dir", 0)
-                .expect("Failed to open directory");
-            
-            if let Some(file) = dir_obj.as_file() {
-                // Read directory entries
-                let mut buffer = [0u8; 512]; // Increased buffer size
-                let mut entries_read = 0;
-                
-                loop {
-                    match file.read(&mut buffer) {
-                        Ok(0) => break, // EOF
-                        Ok(_) => entries_read += 1,
-                        Err(e) => panic!("Failed to read directory entry: {:?}", e),
-                    }
-                }
-                
-                // Should match readdir result (3 entries: ".", "..", "visible_file.txt")
-                assert_eq!(entries_read, 3, "Directory object should read same number of entries as readdir");
-            } else {
-                panic!("Expected FileObject from directory open");
-            }
-        }
-        Err(e) => panic!("Failed to create overlay mount: {:?}", e),
-    }
-}
