+++ kernel/src/fs/mod.rs
@@ -297,131 +297,6 @@ pub trait FileObject: StreamOps {
     }
 }
 
-/// Trait defining basic file system operations
-pub trait FileSystem: Send + Sync {
-    /// Mount operation
-    fn mount(&mut self, mount_point: &str) -> Result<(), FileSystemError>;
-
-    /// Unmount operation
-    fn unmount(&mut self) -> Result<(), FileSystemError>;
-
-    /// Get the name of the file system
-    fn name(&self) -> &str;
-}
-
-/// Trait defining file operations
-pub trait FileOperations: Send + Sync {
-    /// Open a file
-    fn open(&self, path: &str, flags: u32) -> Result<Arc<dyn FileObject>, FileSystemError>;
-
-    /// Read directory entries
-    fn readdir(&self, path: &str) -> Result<Vec<DirectoryEntryInternal>, FileSystemError>;
-    
-    /// Create a file with the specified type.
-    /// 
-    /// # Arguments
-    /// 
-    /// * `path` - The path to the file to create.
-    /// * `file_type` - The type of file to create. This can be a regular file, a device file, or other supported types.
-    /// 
-    /// # Behavior
-    /// 
-    /// - **Regular Files**: These are standard files used for storing data. They are created in the filesystem and can be read from or written to using standard file operations.
-    /// - **Device Files**: These represent hardware devices and are typically used for interacting with device drivers. Creating a device file may involve additional steps, such as associating the file with a specific device driver or hardware resource.
-    /// 
-    /// # Side Effects
-    /// 
-    /// - Creating a device file may require elevated permissions or specific system configurations.
-    /// - If a file already exists at the specified path, the function will return an error of type `FileSystemErrorKind::AlreadyExists`.
-    /// 
-    /// # Returns
-    /// 
-    /// * `Result<(), FileSystemError>` - `Ok` if the file was created successfully, or an error if the operation failed. Errors may include `PermissionDenied`, `InvalidPath`, or `DeviceError` for device files.
-    fn create_file(&self, path: &str, file_type: FileType) -> Result<(), FileSystemError>;
-    
-    /// Create a directory
-    fn create_dir(&self, path: &str) -> Result<(), FileSystemError>;
-
-    /// Remove a file/directory
-    /// 
-    /// This method should handle link count management internally.
-    /// For hardlinks, it decrements link_count and only removes
-    /// actual file data when link_count reaches zero.
-    /// For directories, implementation may require them to be empty.
-    fn remove(&self, path: &str) -> Result<(), FileSystemError>;
-
-    /// Get the metadata
-    fn metadata(&self, path: &str) -> Result<FileMetadata, FileSystemError>;
-
-    /// Truncate a file to the specified size
-    /// 
-    /// This method changes the size of a file to the specified length.
-    /// If the new size is smaller than the current size, the file is truncated.
-    /// If the new size is larger, the file is extended with zero bytes.
-    /// 
-    /// # Arguments
-    /// 
-    /// * `path` - Path to the file to truncate
-    /// * `size` - New size of the file in bytes
-    /// 
-    /// # Returns
-    /// 
-    /// * `Result<(), FileSystemError>` - Ok if the file was truncated successfully
-    /// 
-    /// # Errors
-    /// 
-    /// * `FileSystemError` - If the file doesn't exist, is a directory,
-    ///   or the operation is not supported by this filesystem
-    fn truncate(&self, path: &str, size: u64) -> Result<(), FileSystemError> {
-        let _ = (path, size);
-        Err(FileSystemError {
-            kind: FileSystemErrorKind::NotSupported,
-            message: "Truncate not supported by this filesystem".to_string(),
-        })
-    }
-
-    /// Create a hard link
-    /// 
-    /// Creates a new directory entry that points to the same file data
-    /// as the target. Both entries will have the same file_id and the
-    /// link_count will be incremented.
-    /// 
-    /// # Arguments
-    /// 
-    /// * `target_path` - Path to the existing file to link to
-    /// * `link_path` - Path where the new hard link should be created
-    /// 
-    /// # Returns
-    /// 
-    /// * `Result<(), FileSystemError>` - Ok if the hard link was created successfully
-    /// 
-    /// # Errors
-    /// 
-    /// * `FileSystemError` - If the target doesn't exist, link creation fails,
-    ///   or hard links are not supported by this filesystem
-    fn create_hardlink(&self, target_path: &str, link_path: &str) -> Result<(), FileSystemError> {
-        let _ = (target_path, link_path);
-        Err(FileSystemError {
-            kind: FileSystemErrorKind::NotSupported,
-            message: "Hard links not supported by this filesystem".to_string(),
-        })
-    }
-
-    /// Get the root directory of the file system
-    fn root_dir(&self) -> Result<Directory, FileSystemError> {
-        Err(FileSystemError {
-            kind: FileSystemErrorKind::NotSupported,
-            message: "Root directory not supported by this filesystem".to_string(),
-        })
-    }
-}
-
-/// Trait combining the complete VFS interface
-pub trait VirtualFileSystem: FileSystem + FileOperations {}
-
-// Automatically implement VirtualFileSystem if both FileSystem and FileOperations are implemented
-impl<T: FileSystem + FileOperations> VirtualFileSystem for T {}
-
 /// Enum defining the type of file system
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum FileSystemType {
