--- kernel/src/fs/syscall.rs
@@ -1,516 +1,0 @@
-use alloc::{string::String, vec::Vec, string::ToString, sync::Arc};
-use alloc::boxed::Box;
-
-use crate::{arch::Trapframe, fs::FileType, library::std::string::cstring_to_string, task::mytask};
-
-use super::{SeekFrom, MAX_PATH_LENGTH};
-use super::vfs_v2::manager::VfsManager;
-use super::vfs_v2::core::FileSystemOperations;
-
-fn to_absolute_path_v2(task: &crate::task::Task, path: &str) -> Result<String, ()> {
-    if path.starts_with('/') {
-        Ok(path.to_string())
-    } else {
-        let cwd = task.cwd.clone().ok_or(())?;
-        let mut absolute_path = cwd;
-        if !absolute_path.ends_with('/') {
-            absolute_path.push('/');
-        }
-        absolute_path.push_str(path);
-        // 簡易正規化（"//"や"."などの除去）
-        let mut components = Vec::new();
-        for comp in absolute_path.split('/') {
-            match comp {
-                "" | "." => {},
-                ".." => { components.pop(); },
-                _ => components.push(comp),
-            }
-        }
-        Ok("/".to_string() + &components.join("/"))
-    }
-}
-
-pub fn sys_open(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let path_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let _flags = trapframe.get_arg(1) as i32;
-    let _mode = trapframe.get_arg(2) as i32;
-
-    // Increment PC to avoid infinite loop if open fails
-    trapframe.increment_pc_next(task);
-
-    // Parse path as a null-terminated C string
-    let mut path_bytes = Vec::new();
-    let mut i = 0;
-    unsafe {
-        loop {
-            let byte = *path_ptr.add(i);
-            if byte == 0 {
-                break;
-            }
-            path_bytes.push(byte);
-            i += 1;
-
-            if i > MAX_PATH_LENGTH {
-                return usize::MAX; // Path too long
-            }
-        }
-    }
-
-    // Convert path bytes to string
-    let path_str = match str::from_utf8(&path_bytes) {
-        Ok(s) => s.to_string(),
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-
-    // Try to open the file using VFS
-    let vfs = match task.get_vfs() {
-        Some(vfs) => vfs,
-        None => return usize::MAX, // VFS not initialized
-    };
-    let file_obj = vfs.open(&path_str, 0);
-    match file_obj {
-        Ok(kernel_obj) => {
-            // file_obj is already a KernelObject::File
-            let handle = task.handle_table.insert(kernel_obj);
-            match handle {
-                Ok(handle) => handle as usize,
-                Err(_) => usize::MAX, // Handle table full
-            }
-        }
-        Err(_) => usize::MAX, // File open error
-    }
-}
-
-pub fn sys_close(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as u32; // Handle is u32
-    trapframe.increment_pc_next(task);
-    if task.handle_table.remove(fd).is_some() {
-        0
-    } else {
-        usize::MAX // -1
-    }
-}
-
-pub fn sys_read(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as u32; // Handle is u32
-    let buf_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(1)).unwrap() as *mut u8;
-    let count = trapframe.get_arg(2) as usize;
-
-    // Increment PC to avoid infinite loop if read fails
-    trapframe.increment_pc_next(task);
-
-    let kernel_obj = task.handle_table.get(fd);
-    if kernel_obj.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let kernel_obj = kernel_obj.unwrap();
-    let stream = kernel_obj.as_stream();
-    if stream.is_none() {
-        return usize::MAX; // Object doesn't support stream operations
-    }
-
-    let stream = stream.unwrap();
-    let buffer = unsafe { core::slice::from_raw_parts_mut(buf_ptr, count) };
-    
-    match stream.read(buffer) {
-        Ok(n) => {
-            n
-        }
-        Err(_) => {
-            return usize::MAX; // Read error
-        }
-    }
-}
-
-pub fn sys_write(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as u32; // Handle is u32
-    let buf_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(1)).unwrap() as *const u8;
-    let count = trapframe.get_arg(2) as usize;
-
-    // Increment PC to avoid infinite loop if write fails
-    trapframe.increment_pc_next(task);
-
-    let kernel_obj = task.handle_table.get(fd);
-    if kernel_obj.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let kernel_obj = kernel_obj.unwrap();
-    let stream = kernel_obj.as_stream();
-    if stream.is_none() {
-        return usize::MAX; // Object doesn't support stream operations
-    }
-
-    let stream = stream.unwrap();
-    let buffer = unsafe { core::slice::from_raw_parts(buf_ptr, count) };
-    
-    match stream.write(buffer) {
-        Ok(n) => {
-            n
-        }
-        Err(_) => {
-            return usize::MAX; // Write error
-        }
-    }
-}
-
-pub fn sys_lseek(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as u32; // Handle is u32
-    let offset = trapframe.get_arg(1) as i64;
-    let whence = trapframe.get_arg(2) as i32;
-
-    // Increment PC to avoid infinite loop if lseek fails
-    trapframe.increment_pc_next(task);
-
-    let kernel_obj = task.handle_table.get(fd);
-    if kernel_obj.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let kernel_obj = kernel_obj.unwrap();
-    let file = kernel_obj.as_file();
-    if file.is_none() {
-        return usize::MAX; // Object doesn't support file operations
-    }
-
-    let file = file.unwrap();
-    let whence = match whence {
-        0 => SeekFrom::Start(offset as u64),
-        1 => SeekFrom::Current(offset),
-        2 => SeekFrom::End(offset),
-        _ => return usize::MAX, // Invalid whence
-    };
-
-    match file.seek(whence) {
-        Ok(pos) => {
-            pos as usize
-        }
-        Err(_) => {
-            return usize::MAX; // Lseek error
-        }
-    }
-}
-
-pub fn sys_truncate(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let path_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let length = trapframe.get_arg(1) as u64;
-    
-    trapframe.increment_pc_next(task);
-
-    // Convert path bytes to string
-    let path_str = match cstring_to_string(path_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => match to_absolute_path_v2(&task, &s) {
-            Ok(abs_path) => abs_path,
-            Err(_) => return usize::MAX,
-        },
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-    
-    let vfs = match task.vfs.as_ref() {
-        Some(vfs) => vfs,
-        None => return usize::MAX, // VFS not initialized
-    };
-    
-    match vfs.truncate(&path_str, length) {
-        Ok(_) => 0,
-        Err(_) => usize::MAX, // -1
-    }
-}
-
-pub fn sys_ftruncate(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as u32;
-    let length = trapframe.get_arg(1) as u64;
-    
-    trapframe.increment_pc_next(task);
-    
-    let kernel_obj = task.handle_table.get(fd);
-    if kernel_obj.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-    
-    let kernel_obj = kernel_obj.unwrap();
-    let file = kernel_obj.as_file();
-    if file.is_none() {
-        return usize::MAX; // Object doesn't support file operations
-    }
-    
-    let file = file.unwrap();
-    match file.truncate(length) {
-        Ok(_) => 0,
-        Err(_) => usize::MAX, // -1
-    }
-}
-
-pub fn sys_mkfile(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let path_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let mode = trapframe.get_arg(1) as i32;
-
-    trapframe.increment_pc_next(task);
-
-    // Convert path bytes to string
-    let path_str = match cstring_to_string(path_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => match to_absolute_path_v2(&task, &s) {
-            Ok(abs_path) => abs_path,
-            Err(_) => return usize::MAX,
-        },
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-    
-    let vfs = match task.vfs.as_ref() {
-        Some(vfs) => vfs,
-        None => return usize::MAX, // VFS not initialized
-    };
-
-    match vfs.create_file(&path_str, FileType::RegularFile) {
-        Ok(_) => 0,
-        Err(_) => usize::MAX, // -1
-    }
-}
-
-pub fn sys_mkdir(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let path_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    
-    trapframe.increment_pc_next(task);
-
-    // Convert path bytes to string
-    let path_str = match cstring_to_string(path_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => match to_absolute_path_v2(&task, &s) {
-            Ok(abs_path) => abs_path,
-            Err(_) => return usize::MAX,
-        },
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-    
-    let vfs = match task.vfs.as_ref() {
-        Some(vfs) => vfs,
-        None => return usize::MAX, // VFS not initialized
-    };
-    
-    match vfs.create_dir(&path_str) {
-        Ok(_) => 0,
-        Err(_) => usize::MAX, // -1
-    }
-}
-
-pub fn sys_mount(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let source_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let target_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(1)).unwrap() as *const u8;
-    let fstype_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(2)).unwrap() as *const u8;
-    let flags = trapframe.get_arg(3) as u32;
-    let data_ptr = if trapframe.get_arg(4) == 0 {
-        core::ptr::null()
-    } else {
-        task.vm_manager.translate_vaddr(trapframe.get_arg(4)).unwrap() as *const u8
-    };
-
-    trapframe.increment_pc_next(task);
-
-    // Convert paths and parameters to strings
-    let source_str = match cstring_to_string(source_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => s,
-        Err(_) => return usize::MAX,
-    };
-    
-    let target_str = match cstring_to_string(target_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => s,
-        Err(_) => return usize::MAX,
-    };
-    
-    let fstype_str = match cstring_to_string(fstype_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => s,
-        Err(_) => return usize::MAX,
-    };
-    
-    let data_str = if !data_ptr.is_null() {
-        match cstring_to_string(data_ptr, MAX_PATH_LENGTH) {
-            Ok((s, _)) => Some(s),
-            Err(_) => return usize::MAX,
-        }
-    } else {
-        None
-    };
-
-    // Get VFS reference
-    let vfs = match task.vfs.as_ref() {
-        Some(vfs) => vfs,
-        None => return usize::MAX,
-    };
-
-    // Handle different mount types
-    match fstype_str.as_str() {
-        "bind" => {
-            // Handle bind mount - this is a special case handled by VFS
-            let read_only = (flags & 1) != 0; // MS_RDONLY
-            match vfs.bind_mount(&source_str, &target_str, read_only) {
-                Ok(_) => 0,
-                Err(_) => usize::MAX,
-            }
-        },
-        "overlay" => {
-            // Handle overlay mount - this is a special case handled by VFS
-            if let Some(data) = data_str {
-                match parse_overlay_options(&data) {
-                    Ok((upperdir, lowerdirs)) => {
-                        let lowerdir_refs: Vec<&str> = lowerdirs.iter().map(|s| s.as_str()).collect();
-                        match vfs.overlay_mount(upperdir.as_deref(), lowerdir_refs, &target_str) {
-                            Ok(_) => 0,
-                            Err(_) => usize::MAX,
-                        }
-                    },
-                    Err(_) => usize::MAX,
-                }
-            } else {
-                usize::MAX // Overlay requires options
-            }
-        },
-        _ => {
-            // Handle filesystem creation using drivers
-            let options = data_str.unwrap_or_default();
-            match create_filesystem_and_mount(vfs, &fstype_str, &target_str, &options) {
-                Ok(_) => 0,
-                Err(_) => usize::MAX,
-            }
-        }
-    }
-}
-
-// Helper function to parse overlay mount options
-fn parse_overlay_options(data: &str) -> Result<(Option<String>, Vec<String>), ()> {
-    let mut upperdir = None;
-    let mut lowerdirs = Vec::new();
-    
-    for option in data.split(',') {
-        if let Some(value) = option.strip_prefix("upperdir=") {
-            upperdir = Some(value.to_string());
-        } else if let Some(value) = option.strip_prefix("lowerdir=") {
-            // Multiple lowerdirs can be separated by ':'
-            for lowerdir in value.split(':') {
-                lowerdirs.push(lowerdir.to_string());
-            }
-        }
-    }
-    
-    if lowerdirs.is_empty() {
-        return Err(()); // At least one lowerdir is required
-    }
-    
-    Ok((upperdir, lowerdirs))
-}
-
-/// Create a filesystem using the driver and mount it
-/// 
-/// This function uses the new driver-based approach where option parsing
-/// is delegated to the filesystem driver, and registration is handled
-/// by sys_mount.
-fn create_filesystem_and_mount(
-    vfs: &VfsManager,
-    fstype: &str,
-    target: &str,
-    options: &str,
-) -> Result<(), crate::fs::FileSystemError> {
-    use crate::fs::get_fs_driver_manager;
-    let driver_manager = get_fs_driver_manager();
-    let filesystem: alloc::sync::Arc<dyn FileSystemOperations> = driver_manager.create_from_option_string(fstype, options)?;
-    vfs.mount(filesystem, target, 0)?;
-    Ok(())
-}
-
-pub fn sys_umount(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let target_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let _flags = trapframe.get_arg(1) as u32; // Reserved for future use
-
-    trapframe.increment_pc_next(task);
-
-    // Convert target path to string
-    let target_str = match cstring_to_string(target_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => match to_absolute_path_v2(&task, &s) {
-            Ok(abs_path) => abs_path,
-            Err(_) => return usize::MAX,
-        },
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-
-    // Get VFS reference
-    let vfs = match task.vfs.as_ref() {
-        Some(vfs) => vfs,
-        None => return usize::MAX,
-    };
-
-    // Perform umount operation
-    match vfs.unmount(&target_str) {
-        Ok(_) => 0,
-        Err(_) => usize::MAX,
-    }
-}
-
-pub fn sys_pivot_root(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let new_root_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let old_root_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(1)).unwrap() as *const u8;
-
-    trapframe.increment_pc_next(&task);
-
-    // Convert new_root path to string
-    let new_root_str = match cstring_to_string(new_root_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => match to_absolute_path_v2(&task, &s) {
-            Ok(abs_path) => abs_path,
-            Err(_) => return usize::MAX,
-        },
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-
-    // Convert old_root path to string
-    let old_root_str = match cstring_to_string(old_root_ptr, MAX_PATH_LENGTH) {
-        Ok((s, _)) => match to_absolute_path_v2(&task, &s) {
-            Ok(abs_path) => abs_path,
-            Err(_) => return usize::MAX,
-        },
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-
-    // Get current VFS reference - pivot_root requires isolated VFS namespace
-    let current_vfs = match task.vfs.as_ref() {
-        Some(vfs) => vfs.clone(),
-        None => {
-            // pivot_root requires a task-specific VFS namespace
-            // Tasks without VFS should use the global namespace, but pivot_root
-            // is a namespace operation that doesn't make sense in that context
-            return usize::MAX;
-        },
-    };
-
-    // Perform pivot_root by replacing the mount_tree inside the existing VfsManager
-    match pivot_root_in_place(&current_vfs, &new_root_str, &old_root_str) {
-        Ok(_) => 0,
-        Err(_) => usize::MAX,
-    }
-}
-
-/// Pivot root by replacing the mount tree inside the existing VfsManager
-/// 
-/// This function implements pivot_root without creating a new VfsManager instance.
-/// Instead, it manipulates the mount_tree directly to achieve the same effect.
-/// This approach preserves the relationship between the init process and the global VFS.
-fn pivot_root_in_place(
-    vfs: &Arc<VfsManager>, 
-    new_root_path: &str, 
-    old_root_path: &str
-) -> Result<(), super::FileSystemError> {
-    // TODO: v2用のpivot_root実装が必要
-    Err(super::FileSystemError {
-        kind: super::FileSystemErrorKind::NotSupported,
-        message: "pivot_root_in_place is not yet implemented for VFS v2".to_string(),
-    })
-}
