--- kernel/src/fs/vfs_v2/design_analysis.rs
@@ -1,226 +1,0 @@
-//! Design decision documentation and validation
-//!
-//! This file contains detailed analysis of the VFS v2 reference management design,
-//! specifically addressing the question: "Should children keep their parents alive?"
-
-use alloc::{string::String, vec::Vec, vec};
-
-/// Analysis of different parent-child reference strategies in VFS
-pub struct VfsReferenceAnalysis;
-
-impl VfsReferenceAnalysis {
-    /// Document the specific scenarios where parent-child reference strategy matters
-    pub fn document_scenarios() -> Vec<String> {
-        vec![
-            // Scenario 1: Deep directory traversal
-            "Deep directory traversal (/usr/local/bin/app/config/settings.json)".to_string(),
-            "- Arc parent: Entire path chain stays in memory permanently".to_string(),
-            "- Weak parent: Only actively accessed parts stay cached".to_string(),
-            "".to_string(),
-            
-            // Scenario 2: File operations
-            "File operations on leaf nodes".to_string(),
-            "- Arc parent: Parent directories cannot be evicted while file is open".to_string(),
-            "- Weak parent: Parent path can be reconstructed if needed".to_string(),
-            "".to_string(),
-            
-            // Scenario 3: Temporary files
-            "Temporary file creation/deletion".to_string(),
-            "- Arc parent: Temporary file prevents parent directory cleanup".to_string(),
-            "- Weak parent: Parent cleanup independent of temporary file lifecycle".to_string(),
-            "".to_string(),
-            
-            // Scenario 4: Symbolic links
-            "Symbolic link resolution across multiple directories".to_string(),
-            "- Arc parent: Link target paths keep source paths alive".to_string(),
-            "- Weak parent: Independent cleanup of source and target paths".to_string(),
-        ]
-    }
-
-    /// Memory usage analysis for different designs
-    pub fn memory_usage_analysis() -> String {
-        r#"Memory Usage Analysis:
-
-Arc Parent Design:
-- Memory growth: O(path_depth * accessed_files)
-- Cleanup: Only when no references exist anywhere in subtree
-- Risk: Memory leaks in cyclic mount scenarios
-- Benefit: Guaranteed path consistency
-
-Weak Parent Design (Current):
-- Memory growth: O(actively_accessed_nodes)
-- Cleanup: Immediate when no active references
-- Risk: Temporary path inconsistency
-- Benefit: Optimal memory usage, natural garbage collection
-
-Real-world impact:
-- Large directory trees (e.g., /usr with thousands of files)
-- Arc: Entire tree stays in memory once accessed
-- Weak: Only hot paths stay cached
-"#.to_string()
-    }
-
-    /// Performance characteristics comparison
-    pub fn performance_analysis() -> String {
-        r#"Performance Analysis:
-
-Cache Hit Performance:
-- Arc parent: O(1) - parent always available
-- Weak parent: O(1) - parent available if cached
-- Winner: Tie (both are O(1) for cached access)
-
-Cache Miss Performance:
-- Arc parent: O(1) - parent guaranteed to exist
-- Weak parent: O(depth) - may need path reconstruction
-- Winner: Arc parent (but rare in practice)
-
-Memory Pressure Performance:
-- Arc parent: Poor - cannot free memory under pressure
-- Weak parent: Good - automatic cleanup under pressure
-- Winner: Weak parent (critical for kernel)
-
-Overall System Performance:
-- Arc parent: Good for path-heavy workloads, poor for memory-constrained systems
-- Weak parent: Good for balanced workloads, optimal for kernel environments
-- Winner: Weak parent (better for OS kernel)
-"#.to_string()
-    }
-
-    /// Comparison with other operating systems
-    pub fn os_comparison() -> String {
-        r#"Comparison with Other Operating Systems:
-
-Linux dentry cache:
-- Uses Arc-like strong references for active dentries
-- BUT: Has complex LRU shrinking mechanism
-- Can evict parent dentries under memory pressure
-- Uses RCU for safe concurrent access
-
-FreeBSD namecache:
-- vnodes don't inherently hold parent references
-- Path information reconstructed via namecache lookup
-- Similar to our Weak parent approach
-
-Windows NT object manager:
-- Object references are more complex
-- Parent-child relationships managed separately from object lifetime
-
-Scarlet VFS v2 Design Decision:
-- Weak parent references for simplicity and memory efficiency  
-- No complex LRU shrinking needed
-- Natural garbage collection
-- Reconstruction via filesystem driver lookup
-- Optimal for microkernel/embedded environments
-"#.to_string()
-    }
-
-    /// Final design rationale
-    pub fn design_rationale() -> String {
-        r#"Final Design Rationale:
-
-WHY Weak Parent References:
-
-1. Kernel Memory Constraints:
-   - Kernel heap is limited
-   - Cannot afford memory leaks
-   - Need predictable cleanup behavior
-
-2. Simplicity:
-   - No complex LRU mechanisms needed
-   - No reference counting complications
-   - Clear ownership semantics
-
-3. Filesystem Semantics:
-   - VfsEntry is a CACHE, not persistent storage
-   - Actual file data lives in VfsNode
-   - Path information can be reconstructed
-
-4. Performance Trade-offs:
-   - Slight performance cost on cache miss
-   - BUT: Memory efficiency enables better overall performance
-   - Cache hit performance unaffected
-
-5. Future Extensibility:
-   - Easier to add sophisticated caching later
-   - No breaking changes to core design
-   - Compatible with memory pressure handling
-
-CONCLUSION: Weak parent references provide the best balance of
-simplicity, memory efficiency, and performance for a kernel VFS.
-"#.to_string()
-    }
-}
-
-/// Practical examples demonstrating the design choice
-pub struct PracticalExamples;
-
-impl PracticalExamples {
-    /// Example 1: Build system with many temporary files
-    pub fn build_system_example() -> String {
-        r#"Build System Example:
-
-Scenario: Compilation creates many temporary files in /tmp/build/
-
-Arc Parent Design:
-- /tmp entry stays in memory
-- /tmp/build stays in memory  
-- All intermediate .o files keep build/ alive
-- Memory usage grows linearly with compilation
-- Cleanup only after all files closed
-
-Weak Parent Design:
-- /tmp entry cached only while accessed
-- /tmp/build cached only while files active
-- Individual .o files don't prevent cleanup
-- Memory usage stays constant
-- Natural cleanup as compilation progresses
-
-Result: Weak design provides better memory behavior for build workloads.
-"#.to_string()
-    }
-
-    /// Example 2: Web server serving static files
-    pub fn web_server_example() -> String {
-        r#"Web Server Example:
-
-Scenario: Web server serving files from /var/www/html/
-
-Arc Parent Design:
-- /var, /var/www, /var/www/html permanently cached
-- All parent directories stay in memory
-- Memory usage proportional to directory depth
-- Good for frequently accessed paths
-
-Weak Parent Design:  
-- Only actively served paths stay cached
-- Parent directories cleaned up between requests
-- Memory usage proportional to concurrent requests
-- Path reconstruction on cache miss
-
-Result: Weak design provides better memory efficiency, 
-acceptable performance for typical web workloads.
-"#.to_string()
-    }
-
-    /// Example 3: Database with many temporary tables
-    pub fn database_example() -> String {
-        r#"Database Example:
-
-Scenario: Database creating temporary tables in /var/lib/db/temp/
-
-Arc Parent Design:
-- /var/lib/db/temp stays permanently in memory
-- Each temporary table keeps parent chain alive
-- Memory grows with number of temporary tables
-- Parent cleanup blocked by any remaining table
-
-Weak Parent Design:
-- /var/lib/db/temp cached only when accessed
-- Temporary tables don't prevent parent cleanup
-- Memory usage independent of temp table count
-- Efficient cleanup as tables are dropped
-
-Result: Weak design essential for database workloads with many temp objects.
-"#.to_string()
-    }
-}
