+++ kernel/src/fs/vfs_v2/drivers/initramfs.rs
@@ -1,4 +1,4 @@
-//! Initramfs mounting functionality
+//! Initramfs mounting functionality (VFS v2)
 //!
 //! This module provides functionality to mount the initial ramdisk (initramfs)
 //! as the root filesystem during early kernel boot. The initramfs is loaded by
@@ -8,9 +8,8 @@
 //! at the root ("/") mount point.
 
 use core::ptr;
-
+use alloc::string::ToString;
 use alloc::sync::Arc;
-
 use crate::device::fdt::FdtManager;
 use crate::fs::VfsManager;
 use crate::early_println;
@@ -20,32 +19,15 @@ use crate::vm::vmem::MemoryArea;
 static mut INITRAMFS_AREA: Option<MemoryArea> = None;
 
 /// Relocate initramfs to heap memory
-///
-/// This function copies the initramfs from the location provided by the bootloader
-/// to a new location in kernel heap memory, so that it can be accessed after
-/// virtual memory is enabled.
-///
-/// # Returns
-/// Option<MemoryArea>: The memory area of the relocated initramfs if successful,
-/// None otherwise.
 pub fn relocate_initramfs(usable_area: &mut MemoryArea) -> Result<(), &'static str> {
     early_println!("[InitRamFS] Relocating initramfs to {:#x}", usable_area.start as usize);
-    
-    // Get the FDT manager
     let fdt_manager = FdtManager::get_manager();
-    
-    // Get the initramfs memory area from the device tree
     let original_area = fdt_manager.get_initramfs()
         .ok_or("Failed to get initramfs from device tree")?;
-    
     let size = original_area.size();
-    early_println!("[InitRamFS] Original initramfs at {:#x}, size: {} bytes", 
-        original_area.start, size);
-    
+    early_println!("[InitRamFS] Original initramfs at {:#x}, size: {} bytes", original_area.start, size);
     let new_ptr = usable_area.start as *mut u8;
     usable_area.start = new_ptr as usize + size;
-
-    // Copy the initramfs data
     unsafe {
         ptr::copy_nonoverlapping(
             original_area.start as *const u8,
@@ -53,47 +35,23 @@ pub fn relocate_initramfs(usable_area: &mut MemoryArea) -> Result<(), &'static s
             size
         );
     }
-    
-    // Create a new memory area for the relocated initramfs
     let new_area = MemoryArea::new(new_ptr as usize, (new_ptr as usize) + size - 1);
     early_println!("[InitRamFS] Relocated initramfs to {:#x}", new_area.start);
-    
     unsafe { INITRAMFS_AREA = Some(new_area) };
-
     Ok(())
 }
 
-/// Mount the initramfs as the root filesystem
-///
-/// This function creates a CPIO filesystem from the initramfs memory area
-/// and mounts it at the root ("/") mount point.
-///
-/// # Arguments
-/// * `manager` - A reference to the VFS manager.
-/// * `initramfs` - The memory area of the initramfs.
-///
-/// # Returns
-/// Result<(), FileSystemError>: Ok if mounting was successful, Err otherwise.
 fn mount_initramfs(manager: &Arc<VfsManager>, initramfs: MemoryArea) -> Result<(), FileSystemError> {
     early_println!("[InitRamFS] Initializing initramfs");
-    
-    early_println!("[InitRamFS] Using initramfs at address: {:#x}, size: {} bytes", 
-        initramfs.start, initramfs.size());
-    
-    // Create and register a CPIO filesystem from the initramfs memory area
-    let fs_id = manager.create_and_register_memory_fs("cpiofs", &initramfs)?;
-    
-    // Mount the filesystem at the root directory
-    match manager.mount(fs_id, "/") {
-        Ok(_) => {
-            early_println!("[InitRamFS] Successfully mounted initramfs at root directory");
-            Ok(())
-        },
-        Err(e) => {
-            early_println!("[InitRamFS] Failed to mount initramfs: {:?}", e);
-            Err(e)
-        }
-    }
+    early_println!("[InitRamFS] Using initramfs at address: {:#x}, size: {} bytes", initramfs.start, initramfs.size());
+    // CPIOイメージからファイルシステムを生成
+    let cpio_data = unsafe {
+        core::slice::from_raw_parts(initramfs.start as *const u8, initramfs.size())
+    };
+    let fs = crate::fs::vfs_v2::drivers::cpiofs::CpioFS::new("initramfs".to_string(), cpio_data)?;
+    manager.mount(fs, "/", 0)?;
+    early_println!("[InitRamFS] Successfully mounted initramfs at root directory");
+    Ok(())
 }
 
 #[allow(static_mut_refs)]
@@ -101,8 +59,6 @@ pub fn init_initramfs(manager: &Arc<VfsManager>) {
     let initramfs_ptr = unsafe { INITRAMFS_AREA.as_ref().map(|area| area.start as *const u8).unwrap_or(core::ptr::null()) };
     if !initramfs_ptr.is_null() {
         let initramfs = unsafe { *INITRAMFS_AREA.as_ref().unwrap() };
-        
-        // Mount the initramfs
         if let Err(e) = mount_initramfs(manager, initramfs.clone()) {
             early_println!("[InitRamFS] Warning: Could not mount initramfs: {:?}", e);
             return;
