+++ kernel/src/fs/vfs_v2/drivers/mod.rs
@@ -1,0 +1,4 @@
+pub mod overlayfs;
+pub mod cpiofs;
+pub mod tmpfs;
+pub mod initramfs;
