+++ kernel/src/fs/vfs_v2/mod.rs
@@ -12,29 +12,15 @@
 //! - Cleaner interface for complex filesystems like ext4
 
 pub mod core;
-pub mod tmpfs;
-pub mod cpiofs;
-pub mod overlayfs;
+pub mod drivers;
 pub mod manager;
 pub mod mount_tree;
-pub mod initramfs;
 pub mod syscall;
 
 // VFS v2 test modules
 #[cfg(test)]
 pub mod tests;
-#[cfg(test)]
-pub mod advanced_tests;
-#[cfg(test)]
-pub mod performance_tests;
-#[cfg(test)]
-pub mod cross_vfs_tests;
 
-#[cfg(test)]
-pub mod overlayfs_tests;
 
 
 pub use core::*;
-pub use tmpfs::*;
-pub use cpiofs::*;
-pub use overlayfs::*;
